refactor: search page
